feat: Port StoryState to TypeScript
