+++ src/JObject.ts
@@ -1,0 +1,3 @@
+export interface JObject {
+	[key: string]: any;
+}
